pyproject.toml
[project]
name = "opnote-transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "openai>=2.33.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.49",
]
